change float to double

+++ API/Models/Place.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 #pragma warning disable CA2227 // Collection properties should be read only
 
-/* TODO: Get nearby places
+/* NOTE: Get nearby places
 *  Right now, we're getting the recommended places by state,
 *  If we want be precise, we can use lat and long so we can calculate the distance
 */
@@ -17,10 +17,10 @@ namespace API.Models {
         public string Address { get; set; }
         public string State {get; set;}
         public List<string> Types { get; set; }
-        public float avgOverallRating { get; set; }
-        public float avgCustomerRating { get; set; }
-        public float avgAmentitiesRating { get; set; }
-        public float avgLocationRating { get; set; }
+        public double avgOverallRating { get; set; }
+        public double avgCustomerRating { get; set; }
+        public double avgAmentitiesRating { get; set; }
+        public double avgLocationRating { get; set; }
         public int numOfRatings { get; set; }
     }
 }
